+++ tn40.h
@@ -784,7 +784,7 @@ struct txf_desc {
 u32 bdx_mdio_get(struct bdx_priv *priv);
 int bdx_mdio_write(struct bdx_priv *priv, int device, int port, u16 addr,
 		   u16 data);
-int bdx_mdio_read(struct bdx_priv *priv, int device, int port, u16 addr);
+u16 bdx_mdio_read(struct bdx_priv *priv, int device, int port, u16 addr);
 int bdx_mdio_look_for_phy(struct bdx_priv *priv, int port);
 int bdx_speed_set(struct bdx_priv *priv, u32 speed);
 void bdx_speed_changed(struct bdx_priv *priv, u32 speed);
